+++ Plugin/CustomPosters/src/Data/PosterLayoutProvider.cs
@@ -14,35 +14,55 @@ namespace CustomPosters.Data
 
             if (twoStoryShip)
             {
+                if (shipWindows && widerShip && !Plugin.Service.IsWindow2Enabled)
+                {
+                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship + ShipWindows + WiderShip  (No left window)");
+                    return TwoStoryShip_ShipWindows_WiderShip_NoLeftWindow.Get();
+                }
+
                 if (shipWindows && widerShip)
                 {
                     Plugin.Log.LogInfo("Choosing layout: 2 Story Ship + ShipWindows + WiderShip");
                     return TwoStoryShip_ShipWindows_WiderShip.Get();
                 }
+
+                if (widerShip && !Plugin.Service.EnableLeftWindows)
+                {
+                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship + WiderShip (No left window)");
+                    return TwoStoryShip_WiderShip_NoLeftWindow.Get();
+                }
+
                 if (shipWindows)
                 {
                     Plugin.Log.LogInfo("Choosing layout: 2 Story Ship + ShipWindows");
                     return TwoStoryShip_ShipWindows.Get();
                 }
+
                 if (widerShip)
                 {
                     Plugin.Log.LogInfo("Choosing layout: 2 Story Ship + WiderShip");
                     return TwoStoryShip_WiderShip.Get();
                 }
 
+                if (!Plugin.Service.EnableRightWindows && !Plugin.Service.EnableLeftWindows)
+                {
+                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (No both windows)");
+                    return TwoStoryShip_NoBothWindows.Get();
+                }
+
                 if (Plugin.Service.EnableRightWindows && Plugin.Service.EnableLeftWindows)
                 {
-                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (All Windows)");
+                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (All windows)");
                     return TwoStoryShip_AllWindows.Get();
                 }
                 if (!Plugin.Service.EnableRightWindows)
                 {
-                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (No Right Window)");
+                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (No right window)");
                     return TwoStoryShip_NoRightWindow.Get();
                 }
                 if (!Plugin.Service.EnableLeftWindows)
                 {
-                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (No Left Window)");
+                    Plugin.Log.LogInfo("Choosing layout: 2 Story Ship (No left window)");
                     return TwoStoryShip_NoLeftWindow.Get();
                 }
             }
@@ -55,7 +75,7 @@ namespace CustomPosters.Data
 
             if (widerShip)
             {
-                Plugin.Log.LogInfo($"Choosing layout for WiderShip (Side: {widerShipSide})");
+                Plugin.Log.LogInfo($"Choosing layout: WiderShip - {widerShipSide}");
                 switch (widerShipSide)
                 {
                     case "Both": return WiderShip_Both.Get();