+++ Plugin/CustomPosters/src/Data/TwoStoryShip_NoRightWindow.cs
@@ -7,9 +7,9 @@ namespace CustomPosters.Data.PosterLayouts
         public static PosterData[] Get() => new PosterData[]
         {
             new() { Position = new Vector3(4.0286f, 2.9318f, -16.7774f), Rotation = new Vector3(0, 200.9872f, 0), Scale = new Vector3(0.6391f, 0.4882f, 2f), Name = "Poster1" },
-            new() { Position = new Vector3(9.0884f, 2.4776f, -8.8229f), Rotation = new Vector3(0, 0, 0), Scale = new Vector3(0.7296f, 0.4896f, 1f), Name = "Poster2" },
+            new() { Position = new Vector3(8.8499f, 2.4776f, -8.8229f), Rotation = new Vector3(0, 0, 0), Scale = new Vector3(0.7296f, 0.4896f, 1f), Name = "Poster2" },
             new() { Position = new Vector3(9.9186f, 2.8591f, -17.4716f), Rotation = new Vector3(0, 180f, 356.3345f), Scale = new Vector3(0.7487f, 1.0539f, 1f), Name = "Poster3" },
-            new() { Position = new Vector3(5.3599f, 2.5963f, -9.455f), Rotation = new Vector3(0, 307.2657f, 0), Scale = new Vector3(0.7289f, 0.9989f, 1f), Name = "Poster4" },
+            new() { Position = new Vector3(5.3599f, 2.5963f, -9.5555f), Rotation = new Vector3(0, 308.5657f, 2.68f), Scale = new Vector3(0.7289f, 0.9989f, 1f), Name = "Poster4" },
             new() { Position = new Vector3(5.3282f, 2.7482f, -17.2754f), Rotation = new Vector3(0, 202.3357f, 359.8f), Scale = new Vector3(0.5516f, 0.769f, 1f), Name = "Poster5" },
             new() { Position = new Vector3(2.5679f, 2.6763f, -11.7341f), Rotation = new Vector3(0, 0, 358.6752f), Scale = new Vector3(0.8596f, 1.2194f, 1f), Name = "CustomTips" }
         };