+++ Thunderstore/CHANGELOG.md
@@ -9,12 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added .mp4 support.
   - Added a config option to enable or disable the audio of the poster *mp4* *[ default is false ]* 
-  - Added a config option for each .mp4 poster to change the volume of the video *[ default is false ]*  
+  - Added a config option for each .mp4 poster to change the max volume distance of the video *[ default is 4 ]* 
+  - Added a config option for each .mp4 poster to change the aspect ratio of the video *[ default is 4 ]* 
+  - Added a config option for each .mp4 poster to change the volume of the video *[ default is false ]* 
 
 ### Changed
 - Replaced creating posters via code with AssetBundle.
 - Config overhaul.
 - Refactored and reorganized the entire codebase.
+
+### Fixed
+- Poster positions with *2StoryShipMod*.
 
 ## [2.0.0] - 2025-06-09 - "Major update"
 ### Added